cnct: split resolvers in separate files
